Fix DataType inconsistency in float[] constant buffer handling (#10411)
* Initial plan

* Fix DataType inconsistency in float[] getConstantBuffer method


* Keep allocation tracking using dataType parameter for consistency


---------

+++ nd4j/nd4j-backends/nd4j-backend-impls/nd4j-cpu-backend-common/src/main/java/org/nd4j/linalg/cpu/nativecpu/cache/ConstantBuffersCache.java
@@ -109,16 +109,16 @@ public class ConstantBuffersCache extends BasicConstantHandler {
 
     @Override
     public DataBuffer getConstantBuffer(float[] array, DataType dataType) {
-        ArrayDescriptor descriptor = new ArrayDescriptor(array, DataType.LONG);
+        ArrayDescriptor descriptor = new ArrayDescriptor(array, DataType.FLOAT);
 
         if (!buffersCache.containsKey(descriptor)) {
-            DataBuffer buffer = Nd4j.createTypedBufferDetached(array, DataType.DOUBLE);
+            DataBuffer buffer = Nd4j.createTypedBufferDetached(array, DataType.FLOAT);
 
             if (counter.get() < MAX_ENTRIES) {
                 counter.incrementAndGet();
                 buffersCache.put(descriptor, buffer);
 
-                bytes.addAndGet(array.length * Nd4j.sizeOfDataType(DataType.LONG));
+                bytes.addAndGet(array.length * Nd4j.sizeOfDataType(DataType.FLOAT));
                 AllocationsTracker.getInstance().markAllocated(AllocationKind.CONSTANT, 0, array.length * Nd4j.sizeOfDataType(dataType));
             }
             return buffer;
